+++ tests/Fixtures/Entity/DiscriminatorIntegrityTest/ShortDiscriminatorVehicle.php
@@ -1,0 +1,33 @@
+<?php
+
+/*
+ * This file is part of the Doctrine Doctor.
+ * (c) 2025-2026 Ahmed EBEN HASSINE
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace AhmedBhs\DoctrineDoctor\Tests\Fixtures\Entity\DiscriminatorIntegrityTest;
+
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Entity]
+#[ORM\Table(name: 'disc_integ_vehicles')]
+#[ORM\InheritanceType('SINGLE_TABLE')]
+#[ORM\DiscriminatorColumn(name: 'dtype', type: 'string', length: 8)]
+#[ORM\DiscriminatorMap([
+    'car' => ShortDiscriminatorCar::class,
+    'articulated_lorry' => ShortDiscriminatorLorry::class,
+])]
+class ShortDiscriminatorVehicle
+{
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
+
+    #[ORM\Column(length: 100)]
+    private string $label = '';
+}